refactor: move RealtimeClientOptions to general options file
