+++ packages/supabase/lib/supabase.dart
@@ -11,7 +11,6 @@ export 'package:realtime_client/realtime_client.dart';
 export 'package:storage_client/storage_client.dart';
 
 export 'src/auth_user.dart';
-export 'src/realtime_client_options.dart';
 export 'src/remove_subscription_result.dart';
 export 'src/supabase_client.dart';
 export 'src/supabase_client_options.dart';
